New CSS class typeid for type identifiers.
Standard types are in both the vhdlieee and typeid classes.  Type identifiers
declared at the top level of archtiectures, packages and package bodies are
only in the typeid class.  They are marked globally, without taking scope into
account.  Runtime has further increased.  The POD documentation is not up to
date yet.

A colouration entry has been added to the experimental CSS for the new class,
for testing.