Proper program termination on fatal errors when parallel parsing.
parsefile() used to call croak() (which calls exit) when it could not open a
file, with the effect that only its parsing task terminated.  Now an error is
returned in the "fatal" field of the source file hash in this situation.  The
renamed endtasks() function is called on program termination via END{}, sends
the suicide command to the parsing tasks and waits for them to terminate.