New option --packagesummary to print summary of package declarations.
For now the summary is only printed if there is a type, const, component or
procedure/function declaration which has a @brief description or a description
longer than a certain threshold, and only those kinds of declarations are
listed.  The summary entries link to the full descriptions.  The newly
necessary labels for constants and functions/procedures are generated when
printing the summary.  Functions and procedures were removed from the %labels
prefix hash, as they are not needed without --packagesummary, and their labels
have to reflect polymorphic functions appropriately (special case).

Headline of function/procedure declarations now always "Functions and
procedures" to make clear that there is no other location for either of them.