@see for declarations in packages.
Implemented see-also references for constants, types, components, procedures
and functions.  The references may be declarations in the same package,
selected names of declarations in different packages, or design elements.  The
selected name lookup tables are used to find design elements and packages, and
package declaration arrays are searched for in-package declarations.

For now, @see references in design elements can still only refer to other
design elements.