Allow specifying a port/generic name from which to inherit docs.
An entity name also has to be specified, followed by the port or generic name
separated by a dot.  Currently ports can inherit from generics and vice versa,
which may give rise to ambiguities if there are ports and generics with the
same name; but this at least very unusual if not illegal.