Changes to finding available number of CPU cores for parallel parsing.
The number of cores has to be determined for setting the number of parallel
jobs when this is not explicitly set with -j.  This is now restricted to linux
as it was reported to fail and cause trouble on OS X.  Also, warnings due to
failed `...` are now suppressed; and nproc (which outputs only the number we
want) is tried before taskset (whose output has to be postprocessed).