Parsing of case..generate statements
Parser understands case..generate and corresponding when..=> clauses.
Internally the "when" clauses are represented as siblings rather than children
of the "case..generate", like the elsif and else branches of "if..generate".
The hash key referencing the primary generate statement has been changed to
"parent" for both, to allow analogous treatment in doc generation later.

Small correction to discarding of inner begin block of VHDL 93 "if..generate"
with declarative part - change $parent and $prevlevel consistently with stack.