Parse types in generic lists.
The types are documented by the same code that documents generics.  The
identifier is put first, like for ordinary generic constants and unlike in the
VHDL source.  But this seems more consistent and readable as the "kind" of
generic always comes second.  This order will also be applied to packages and
subprogram generics later.

Also, cleaned up readportlist():  Flags for type of entries parsed; and more
generic variable names.