Parse and document generic parameters of packages, and their instances
Generic parameter lists of generic packages are now parsed and documented.
Instances of generic packages are resolved and documented on the page of the
generic package.

The internal representation of generic maps in for..use statements (also used
for generic packages) was changed to an array of pairs rather than a
sub-declaration hash.

The handling of generic package instances is still incomplete; for example
their own (empty) doc page is still generated.