Generic subprograms (VHDL2008) and associated changes
Parse extended to accept generic lists for subprograms and the "parameter"
keyword.  Instances of generic subprograms are also parsed.  Generic lists are
documented with the generic subprogram; instances are documented with generic
associations where they occur and link to the generic declaration.

The resolve_types() function was refactored and the resolution part reused for
resolving generic subprograms in the new resolve_genfunc() function.  Generic
functions and procedures are now added to the selected names table.  (Local
generic functions in architectures and package bodies need not be added because
they are resolved independently.)