Export generic subprograms from package bodies without package, not components
A small change in handling declarations in orphaned package bodies (i.e.
without matching packages):  Component declarations are not any more added to
the global selected name table, as those that are in the missing package would
not be duplicated in the package body.  Instead generic functions and
procedures are added, because their definition in the package body is likely to
match a declaration in the package.