Try top-level design units first when resolving cross-references
The change applies to resolving selected names without a type specified.  This
is currently only used for "see also" cross-references.  Previously the result
was random in case of ambiguities because of the unpredictable order of hash
keys returned by keys().  This led to references to entities sometimes being
resolved to their component declaration.