Fixes for git substitution in @version etc.
Fix for determining last-change commit of a VHDL file.  Using (?:) to decide
between two regular expressions does not work.

"git describe" is now run only if there are files that need it for @version.

The git hash is substituted for @id only if the file has been found in a
repository commit.  Otherwise files are treated as if they were not in a
repository even when they are in a subdirectory of one.