Changed order of links in HTML page headline
The reasoning is that the hierarchy is the most important, as its top modules
were specifically designated, and should come first.  Next is the package list
because packages often wrap the main contents of design projects that do not
have designated top modules.  Next are the entity list (with
architectures/configurations) and instantiation overview, which belong
together.  Last is the list of sources, which should only be needed as a last
resort when the doc output is not sufficient.