Fixed for LaTeX output: column alignment in lists of ports/generics/constants
The port list output function now takes an indentation argument which is an
integer denoting units of \texindent.  Using hanging paragraph indentation did
not work for tables.

The padding at the start of each table line was removed from all tables by
adding the @{} incantation at the start of the table format.

The hack for the port direction column was replaced by a proper additional
centred column.  The padding after that column is disabled in the table format,
and added explicitly if the items listed have a direction.  So there is no
additional padding for generics and constants, which have no direction.