Don't report markup codes without required arguments as unknown.
The Regexes recognising some markup codes have been made slightly more tolerant
to match even when required arguments are missing.  Assignment to the
documentation hash is suppressed if the arguments are not present (in some
cases, an empty entry was added previously).  This prevents inapproriate
messages about unknown markup codes in the last else branch.