Fix for handling duplicate declarations in file declaration list
The list of declarations in each file that was added to the source overview
page recently is the only place where duplicate declarations occur.  These are
otherwise discarded early in the processing.  Therefore we have to deduplicate
the list by non-selected name and create selected names for output as needed.