doxyconvert: support \file markup; and verbatim blocks at VHDL statements
Doxygen's \file markup code is now supported by the conversion program, with
restrictions:  It is only recognised if it is the very first doc comment in the
file, and is always taken to refer to the current file.  The file documentation
is appended to that of all design units (major top-level declarations) in the
file.

Verbatim blocks now end at VHDL statements, where vhdocl also detects the end
of a doc comment block.  Previously such blocks were ended at lines not
containing any doc comments, which was simpler but makes little sense.